refactor: re-organize interface and types
- create content interface
- stub out reducer to include content
