+++ src/interfaces/IContent.ts
@@ -1,0 +1,22 @@
+export interface IContent {
+  header: {
+    content: object;
+    background: string;
+  };
+  column1: {
+    content: object;
+    background: string;
+  };
+  column2: {
+    content: object;
+    background: string;
+  };
+  column3: {
+    content: object;
+    background: string;
+  };
+  footer: {
+    content: object;
+    background: string;
+  };
+}
