+++ src/interfaces/IWebsite.ts
@@ -1,0 +1,6 @@
+export interface IWebsite {
+  [x: string]: any;
+  id: number;
+  layout: string;
+  title: string;
+}
